[fix] 플레이어 경험치바 버그 수정
경험치바가 표시 안되는 버그 수정

+++ Assets/2. Scripts/UI/PlayerExp.cs
@@ -36,6 +36,6 @@ public class PlayerExp : MonoBehaviour
 
     void ChangeExpBar(BigInteger currentExp, BigInteger maxExp)
     {
-        _slider.value = (float)(currentExp / maxExp);
+        _slider.value = (float)((double)currentExp / (double)maxExp);
     }
 }
